Move counter functions from util.py to counter.py
There was a circular dependency between protocol.py and util.py.
This commit breaks it by moving all counters code from both files to
counter.py.
Works towards #130.
