+++ privcount/protocol.py
@@ -9,6 +9,7 @@ from twisted.protocols.basic import LineOnlyReceiver
 
 from cryptography.hazmat.primitives.hashes import SHA256
 from privcount.util import CryptoHash, get_hmac, verify_hmac, b64_padded_length
+from privcount.counter import get_events_for_counters, get_valid_events
 
 class PrivCountProtocol(LineOnlyReceiver):
     '''
