+++ privcount/share_keeper.py
@@ -6,6 +6,7 @@ Created on Dec 12, 2015
 import os
 import logging
 import cPickle as pickle
+import yaml
 
 from copy import deepcopy
 
@@ -13,9 +14,8 @@ from twisted.internet import reactor, ssl
 from twisted.internet.protocol import ReconnectingClientFactory
 
 from protocol import PrivCountClientProtocol
-from util import SecureCounters, log_error, get_public_digest, generate_keypair, get_serialized_public_key, load_private_key_file, decrypt, normalise_path, counter_modulus, add_counter_limits_to_config, check_noise_weight_config, check_counters_config, combine_counters, choose_secret_handshake_path, PrivCountClient
-
-import yaml
+from counter import SecureCounters, counter_modulus, add_counter_limits_to_config, combine_counters
+from util import log_error, get_public_digest, generate_keypair, get_serialized_public_key, load_private_key_file, decrypt, normalise_path, choose_secret_handshake_path, PrivCountClient
 
 class ShareKeeper(ReconnectingClientFactory, PrivCountClient):
     '''
